requirements.txt
mysql-connector-python==9.1.0
parse==1.20.2
tinytag==2.1.0
